fix(build): Use development profiles to allow local building
