+++ lib/core/system.dart
@@ -1,6 +1,5 @@
 export 'system/configuration.dart';
 export 'system/desktop_window.dart';
-export 'system/environment.dart';
 export 'system/filesystem.dart';
 export 'system/flavor.dart';
 export 'system/in_app_purchases.dart';
