+++ lib/core/system/flavor.dart
@@ -1,7 +1,11 @@
 import 'package:envify/envify.dart';
-import 'package:lunasea/core.dart';
 
 part 'flavor.g.dart';
+
+const _DEVELOP = 'develop';
+const _ALPHA = 'alpha';
+const _BETA = 'beta';
+const _PRODUCTION = 'production';
 
 @Envify(path: '.flavor')
 abstract class Flavor {
@@ -21,5 +25,41 @@ extension FlavorExtension on Flavor {
         return Environment.PRODUCTION;
     }
     throw Exception('Invalid Environment');
+  }
+}
+
+enum Environment {
+  DEVELOP,
+  ALPHA,
+  BETA,
+  PRODUCTION,
+}
+
+extension EnvironmentExtension on Environment {
+  String get key {
+    switch (this) {
+      case Environment.DEVELOP:
+        return _DEVELOP;
+      case Environment.ALPHA:
+        return _ALPHA;
+      case Environment.BETA:
+        return _BETA;
+      case Environment.PRODUCTION:
+        return _PRODUCTION;
+    }
+  }
+
+  Environment from(String key) {
+    switch (key) {
+      case _DEVELOP:
+        return Environment.DEVELOP;
+      case _ALPHA:
+        return Environment.ALPHA;
+      case _BETA:
+        return Environment.BETA;
+      case _PRODUCTION:
+        return Environment.PRODUCTION;
+    }
+    throw Exception('Unknown Flavor');
   }
 }
